[PM-34897] refactor: Extract PremiumFeaturesList as shared Billing component
