+++ BitwardenShared/UI/Billing/PremiumUpgrade/PremiumUpgradeView.swift
@@ -63,7 +63,7 @@ struct PremiumUpgradeView: View {
                 .foregroundColor(Color(asset: SharedAsset.Colors.textSecondary))
                 .padding(.bottom, 16)
 
-            featuresList
+            PremiumFeaturesList()
         }
         .padding(.top, 16)
         .padding(.horizontal, 16)
@@ -81,20 +81,6 @@ struct PremiumUpgradeView: View {
             Text(Localizations.perMonth)
                 .styleGuide(.body)
                 .foregroundColor(Color(asset: SharedAsset.Colors.textSecondary))
-        }
-    }
-
-    /// The list of premium features.
-    private var featuresList: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Divider()
-
-            ContentBlock {
-                featureRow(Localizations.builtInAuthenticator)
-                featureRow(Localizations.emergencyAccess)
-                featureRow(Localizations.secureFileStorage)
-                featureRow(Localizations.breachMonitoring)
-            }
         }
     }
 
@@ -123,25 +109,6 @@ struct PremiumUpgradeView: View {
             .frame(maxWidth: .infinity)
     }
 
-    // MARK: Private Methods
-
-    /// A single feature row with a check icon.
-    ///
-    /// - Parameter text: The feature text to display.
-    /// - Returns: A view displaying the feature with a check icon.
-    ///
-    private func featureRow(_ text: String) -> some View {
-        HStack(spacing: 12) {
-            SharedAsset.Icons.checkCircle24.swiftUIImage
-                .foregroundColor(Color(asset: SharedAsset.Colors.textInteraction))
-                .accessibilityHidden(true)
-
-            Text(text)
-                .styleGuide(.headline, weight: .semibold)
-                .foregroundColor(Color(asset: SharedAsset.Colors.textPrimary))
-        }
-        .padding(.vertical, 16)
-    }
 }
 
 // MARK: - Previews
